feature: Added `appGetStaticProps`
