+++ pages/examples/fetch-global-data.js
@@ -1,6 +1,9 @@
 import AllPosts from '../../components/examples/AllPosts';
-import { getApolloClient, getPosts } from '@wpengine/headless';
-import { gql, useQuery } from '@apollo/client';
+// import { getApolloClient, getPosts } from '@wpengine/headless';
+// import { gql, useQuery } from '@apollo/client';
+
+import appGetStaticProps from "../../lib/api/appGetStaticProps";
+
 /**
  * Example getting with usePosts SSR.
  *
@@ -13,41 +16,27 @@ import { gql, useQuery } from '@apollo/client';
  * - Video explaining this: https://www.loom.com/share/dc720e6d4dd14ef38a011b7a995ad56a
  */
 
-const MENU_QUERY = gql`
-  query {
-    menu(id: "header", idType: NAME) {
-      menuItems {
-        nodes {
-          label
-        }
-      }
-    }
-  }
-`;
 const Blog = (props) => {
-  console.log('Blog props', props);
+    console.log('Blog props', props);
 
-  return (
-    <div>
-      <h1>AllPosts example with global data</h1>
-      <AllPosts />
-    </div>
-  );
+    return (
+        <div>
+            <h1>AllPosts example with global data</h1>
+            <AllPosts />
+        </div>
+    );
 };
 
 export const getStaticProps = async (context) => {
-  const client = getApolloClient(context);
-  const posts = await getPosts(client);
-  const menuData = await client.query({
-    query: MENU_QUERY,
-  });
 
-  return {
-    props: {
-      posts,
-      menuData,
-    },
-  };
+    const otherProps = {
+        properties: ['fakeProperty1', 'fakeProperty2']
+    }
+    const props = await appGetStaticProps(context, otherProps);
+    console.log("props: ", props)
+    return {
+        props
+    }
 };
 
 export default Blog;
